+++ crates/compiler/src/time.rs
@@ -65,7 +65,7 @@ impl TickCounter {
         Self { value }
     }
 
-    pub fn value(&self) -> u32 {
+    pub const fn value(&self) -> u32 {
         self.value
     }