+++ docs/changelog.md
@@ -20,6 +20,7 @@ Audio driver changes:
  * Optimised `skip_last_loop` bytecode instructions
  * Added sound effect dropout behaviour
     * Low-priority sound effects will not play if both sfx channels are active
+    * Added interruptible flag to sound effects
 
 IO Command changes:
  * Removed `SET_ENABLED_CHANNELS` IO command, replaced with `SET_MUSIC_CHANNELS`.